Update the composer packages
Updating the package is necessary in order to be able to install them
via the latest version of composer (1.7.2). With the previously installed
package we got an "Imcompatible declaration" conflict between Symfony flex
and composer.

https://github.com/symfony/flex/issues/397

As a bonus I tried to use as many stable packages as possible. The
default installed package is now also set to stable.