Changed backend Docker setup
Replaced the 2 separate images: apache + php with 1 image: apache with mod php.
This can be split up into separate images later. But this is an easy way to get
everything up and running.

The Dockerfile is also restructured to make it easy to build the image and
fast when only code is changed. The image now contains all code, to make
the application portable.

The cli container is removed because it's possible to use the backend container
to do these kind of actions. Creating a new container:
>\# docker run --rm -ti NAME_OF_DOCKER_IMAGE bin/console

Or reusing an already started container:
>\# docker exec -ti NAME_OF_DOCKER_CONTAINER bin/console

Or if you want to login to the container and use bash;
>\# docker exec -ti NAME_OF_DOCKER_CONTAINER /bin/bash